refactor: move xrmemreader into odc.loader._zarr
